feat: Add search by text and debounce
